separate dax dialog view controller code out
