+++ DuckDuckGo/DaxOnboardingViewController.swift
@@ -85,7 +85,7 @@ class DaxOnboardingViewController: UIViewController, Onboarding {
     
     func showDaxDialog(completion: @escaping () -> Void) {
         let message = "The Internet can be kinda creepy.\n\nNot to worry! Searching and browsing privately is easier than you think."
-        let cta = "Let's Do It"
+        let cta = "Let's Do It!"
         
         daxDialogContainer.alpha = 0.0
         daxDialogContainer.isHidden = false
@@ -98,89 +98,6 @@ class DaxOnboardingViewController: UIViewController, Onboarding {
         }, completion: { _ in
             completion()
         })
-    }
-    
-}
-
-class DaxDialogViewController: UIViewController {
-    
-    @IBOutlet weak var icon: UIView!
-    @IBOutlet weak var pointer: UIView!
-    @IBOutlet weak var textArea: UIView!
-    @IBOutlet weak var button: UIButton!
-    @IBOutlet weak var label: UILabel!
-    
-    var message: String? {
-        didSet {
-            label.text = nil
-            chars = Array(message ?? "")
-        }
-    }
-    
-    var cta: String? {
-        didSet {
-            if let title = cta {
-                button.setTitle(title, for: .normal)
-            } else {
-                button.isHidden = true
-            }
-        }
-    }
-    
-    var onTapCta: (() -> Void)?
-    
-    private var position: Int = 0
-    private var chars = [Character]()
-    
-    private lazy var paragraphStyle: NSParagraphStyle = {
-        let paragraphStyle = NSMutableParagraphStyle()
-        paragraphStyle.lineHeightMultiple = 1.17
-        return paragraphStyle
-    }()
-    
-    private var atEnd: Bool {
-        return position >= chars.count
-    }
-    
-    override func viewDidLoad() {
-        super.viewDidLoad()
-        applyPointerRotation()
-        textArea.displayDropShadow()
-        button.displayDropShadow()
-    }
-    
-    func applyPointerRotation() {
-        let rads = CGFloat(45 * Double.pi / 180)
-        pointer.layer.transform = CATransform3DMakeRotation(rads, 0.0, 0.0, 1.0)
-    }
-    
-    func start() {
-        position = 0
-        showNextChar()
-    }
-    
-    @IBAction func onTapText() {
-        position = chars.count
-        updateMessage()
-    }
-    
-    private func showNextChar() {
-        guard !atEnd else { return }
-        
-        position += 1
-        while !atEnd && self.chars[position].isWhitespace {
-            position += 1
-        }
-        updateMessage()
-        
-        DispatchQueue.main.asyncAfter(deadline: .now() + 0.02) {
-            self.showNextChar()
-        }
-    }
-
-    private func updateMessage() {
-        let message = String(chars[0 ..< position])
-        label.attributedText = NSMutableAttributedString(string: message, attributes: [NSAttributedString.Key.paragraphStyle: paragraphStyle])
     }
     
 }
